+++ CriticalMass/ChatMessage.swift
@@ -1,0 +1,13 @@
+//
+//  ChatMessage.swift
+//  CriticalMaps
+//
+//  Created by Leonard Thomas on 12/17/18.
+//
+
+import Foundation
+
+public struct ChatMessage: Codable, Equatable {
+    var message: String
+    var timestamp: TimeInterval
+}